requirements.txt
streamlit>=1.52.0
pandas>=2.3.0
requests>=2.32.0
pytest>=9.0.0
pyarrow>=22.0.0
aeon>=0.11.0
numpy>=1.24.0
scikit-learn>=1.3.0
pyyaml>=6.0.0
plotly>=5.18.0
